refactor: create packages for each state
