+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/Chat.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright (C) 2023 Sonar Contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package xyz.jonesdev.sonar.common.fallback.protocol.packets.play;
+
+import io.netty.buffer.ByteBuf;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.serializer.json.JSONComponentSerializer;
+import org.jetbrains.annotations.NotNull;
+import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
+import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
+
+import java.util.UUID;
+
+import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.*;
+import static xyz.jonesdev.sonar.common.utility.protocol.ProtocolUtil.writeString;
+import static xyz.jonesdev.sonar.common.utility.protocol.ProtocolUtil.writeUUID;
+import static xyz.jonesdev.sonar.common.utility.protocol.VarIntUtil.writeVarInt;
+
+@Getter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public final class Chat implements FallbackPacket {
+  private static final UUID PLACEHOLDER_UUID = new UUID(0L, 0L);
+  private Component component;
+  private byte position;
+
+  @Override
+  public void encode(final ByteBuf byteBuf, final @NotNull ProtocolVersion protocolVersion) {
+    final String serialized = JSONComponentSerializer.json().serialize(component);
+    writeString(byteBuf, serialized);
+
+    // Message type
+    if (protocolVersion.compareTo(MINECRAFT_1_19_1) >= 0) {
+      byteBuf.writeBoolean(position == 2);
+    } else if (protocolVersion.compareTo(MINECRAFT_1_19) >= 0) {
+      writeVarInt(byteBuf, position);
+    } else {
+      byteBuf.writeByte(this.position);
+    }
+
+    // UUID
+    if (protocolVersion.compareTo(MINECRAFT_1_16) >= 0
+      && protocolVersion.compareTo(MINECRAFT_1_19) < 0) {
+      writeUUID(byteBuf, PLACEHOLDER_UUID);
+    }
+  }
+
+  @Override
+  public void decode(final ByteBuf byteBuf, final ProtocolVersion protocolVersion) {
+    throw new UnsupportedOperationException();
+  }
+}
