+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/TeleportConfirm.java
@@ -15,36 +15,32 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package xyz.jonesdev.sonar.common.fallback.protocol.packets;
+package xyz.jonesdev.sonar.common.fallback.protocol.packets.play;
 
 import io.netty.buffer.ByteBuf;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import org.jetbrains.annotations.NotNull;
 import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
 import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
+
+import static xyz.jonesdev.sonar.common.utility.protocol.VarIntUtil.readVarInt;
 
 @Getter
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class Abilities implements FallbackPacket {
-  private byte encodedFlags;
-  private float flySpeed, walkSpeed;
+public final class TeleportConfirm implements FallbackPacket {
+  private int teleportId;
 
   @Override
-  public void encode(final @NotNull ByteBuf byteBuf, final ProtocolVersion protocolVersion) {
-    byteBuf.writeByte(encodedFlags);
-    byteBuf.writeFloat(flySpeed);
-    byteBuf.writeFloat(walkSpeed);
+  public void encode(final ByteBuf byteBuf, final ProtocolVersion protocolVersion) {
+    throw new UnsupportedOperationException();
   }
 
   @Override
-  public void decode(final @NotNull ByteBuf byteBuf, final ProtocolVersion protocolVersion) {
-    encodedFlags = byteBuf.readByte();
-    flySpeed = byteBuf.readFloat();
-    walkSpeed = byteBuf.readFloat();
+  public void decode(final ByteBuf byteBuf, final ProtocolVersion protocolVersion) {
+    teleportId = readVarInt(byteBuf);
   }
 }
