+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/packets/play/Transaction.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package xyz.jonesdev.sonar.common.fallback.protocol.packets;
+package xyz.jonesdev.sonar.common.fallback.protocol.packets.play;
 
 import io.netty.buffer.ByteBuf;
 import lombok.AllArgsConstructor;
@@ -26,37 +26,40 @@ import org.jetbrains.annotations.NotNull;
 import xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion;
 import xyz.jonesdev.sonar.common.fallback.protocol.FallbackPacket;
 
-import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.MINECRAFT_1_12_2;
-import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.MINECRAFT_1_8;
-import static xyz.jonesdev.sonar.common.utility.protocol.VarIntUtil.readVarInt;
-import static xyz.jonesdev.sonar.common.utility.protocol.VarIntUtil.writeVarInt;
+import static xyz.jonesdev.sonar.api.fallback.protocol.ProtocolVersion.MINECRAFT_1_17;
 
 @Getter
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public final class KeepAlive implements FallbackPacket {
-  private long id;
+public final class Transaction implements FallbackPacket {
+  private int windowId, id;
+  private boolean accepted;
 
   @Override
   public void encode(final ByteBuf byteBuf, final @NotNull ProtocolVersion protocolVersion) {
-    if (protocolVersion.compareTo(MINECRAFT_1_12_2) >= 0) {
-      byteBuf.writeLong(id);
-    } else if (protocolVersion.compareTo(MINECRAFT_1_8) >= 0) {
-      writeVarInt(byteBuf, (int) id);
+    if (protocolVersion.compareTo(MINECRAFT_1_17) < 0) {
+      byteBuf.writeByte(windowId);
+      byteBuf.writeShort((short) id);
+      // The "accepted" field is actually really unnecessary since
+      // it's never even used in the client.
+      byteBuf.writeBoolean(accepted);
     } else {
-      byteBuf.writeInt((int) id);
+      byteBuf.writeInt(id);
     }
   }
 
   @Override
   public void decode(final ByteBuf byteBuf, final @NotNull ProtocolVersion protocolVersion) {
-    if (protocolVersion.compareTo(MINECRAFT_1_12_2) >= 0) {
-      id = byteBuf.readLong();
-    } else if (protocolVersion.compareTo(MINECRAFT_1_8) >= 0) {
-      id = readVarInt(byteBuf);
+    if (protocolVersion.compareTo(MINECRAFT_1_17) < 0) {
+      windowId = byteBuf.readByte();
+      id = byteBuf.readShort();
+      accepted = byteBuf.readBoolean();
     } else {
       id = byteBuf.readInt();
+      // Always set accepted to true since 1.17 or higher don't use
+      // transactions for inventory confirmation anymore.
+      accepted = true;
     }
   }
 }
